CHORE: Move Coordinate modules application -> common.coordinate
